// src/infrastructure/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "10.2.0 (e07807d)"
  }
  public: {
    Tables: {
// ...
      expenses: {
        Row: {
          amount: number | null
          category: string | null
          comment: string | null
          corresponding_product: string | null
          corresponding_qty: number | null
          created_at: string
          department: string | null
          id: number
          type: string | null
        }
        Insert: {
          amount?: number | null
          category?: string | null
          comment?: string | null
          corresponding_product?: string | null
          corresponding_qty?: number | null
          created_at?: string
          department?: string | null
          id?: number
          type?: string | null
        }
        Update: {
          amount?: number | null
          category?: string | null
          comment?: string | null
          corresponding_product?: string | null
          corresponding_qty?: number | null
          created_at?: string
          department?: string | null
          id?: number
          type?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}